[release,cmake] Fixes non-posix sed statement and problems with configure consisting of one unquoted pair.
\S\s seem not to be supported on some systems. Therefore we switch to listing the whitespace
characters explicitly. I addition checking wether the result of sed is the same as the input
is not sufficient for checking whether there was no match. With just one key, unquoted value
pair in the string they are the same even for a match therefore we use grep to only get
matches.