[bugfix,parallel] Do not send or receive message of size zero.
There is not need to send message that contain no data. It might even
produce problems on  some implementation. Therfore this patch always
checks the size and simply sets the request to inactive (aka
MPI_REQUEST_NULL) if no data would be send or received.

The patch is to be cherry-picked to master if Christian approves.