[parallel] Allow users to deactivate detection of MPI implementation.
If users are using MPI implementation that we did not think of, then
configure fails with the error:
configure: error: Could not identify MPI-package! Please send a
bugreport and tell us what MPI-package you're using.

This patch allows the user to deactive this check. Therefore he will
be able to use DUNE as long as he provides the correct CXXFLAGS and
LDFLAGS needed by MPI.